Use "all versions" DOI on README badge
Zenodo stores all releases made from a GitHub repository and assigns
a DOI for each one.  This allows users to always access the specific
version of the code files.

The DOI for v20.0.0 is https://doi.org/10.5281/zenodo.13284703

This commit sets the DOI to https://doi.org/10.5281/zenodo.13284702,
which is a special DOI that refers to "all versions" of the project.
This ensures that users will always find the latest version and does not
need to be changed every time that a new release is made.

For citations, either the specific DOI for the release used can be
cited, or the general DOI with the version number also specified.